refactor(storage): restructure MoveObject transport as Mover helper
