+++ src/storage/src/storage/move_object.rs
@@ -348,3 +348,74 @@ where
         self.stub.move_object(self.request, self.options).await
     }
 }
+
+pub(crate) struct Mover {
+    pub(crate) inner: std::sync::Arc<crate::storage::client::StorageInner>,
+    pub(crate) request: crate::model::MoveObjectRequest,
+    pub(crate) options: crate::storage::request_options::RequestOptions,
+}
+
+impl Mover {
+    pub(crate) async fn response(self) -> Result<Object> {
+        use gaxi::{
+            grpc::tonic::{Extensions, GrpcMethod},
+            prost::ToProto,
+        };
+        let options =
+            google_cloud_gax::options::internal::set_default_idempotency(self.options.gax(), false);
+        let extensions = {
+            let mut e = Extensions::new();
+            e.insert(GrpcMethod::new("google.storage.v2.Storage", "MoveObject"));
+            e
+        };
+        let path = http::uri::PathAndQuery::from_static("/google.storage.v2.Storage/MoveObject");
+        let x_goog_request_params = {
+            use gaxi::routing_parameter::Segment;
+            gaxi::routing_parameter::format(&[None
+                .or_else(|| {
+                    gaxi::routing_parameter::value(
+                        Some(&self.request).map(|m| &m.bucket).map(|s| s.as_str()),
+                        &[],
+                        &[Segment::MultiWildcard],
+                        &[],
+                    )
+                })
+                .map(|v| ("bucket", v))])
+        };
+        if x_goog_request_params.is_empty() {
+            use gaxi::path_parameter::PathMismatchBuilder;
+            use gaxi::routing_parameter::Segment;
+            use google_cloud_gax::error::binding::BindingError;
+            let mut paths = Vec::new();
+            {
+                let builder = PathMismatchBuilder::default();
+                let builder = builder.maybe_add(
+                    Some(&self.request).map(|m| &m.bucket).map(|s| s.as_str()),
+                    &[Segment::MultiWildcard],
+                    "bucket",
+                    "**",
+                );
+                paths.push(builder.build());
+            }
+            return Err(google_cloud_gax::error::Error::binding(BindingError {
+                paths,
+            }));
+        }
+
+        type TR = crate::google::storage::v2::Object;
+        let response = self
+            .inner
+            .grpc
+            .execute(
+                extensions,
+                path,
+                self.request.to_proto().map_err(crate::Error::deser)?,
+                options,
+                &crate::storage::info::X_GOOG_API_CLIENT_HEADER,
+                &x_goog_request_params,
+            )
+            .await
+            .and_then(gaxi::grpc::to_gax_response::<TR, Object>)?;
+        Ok(response.into_body())
+    }
+}
